+++ src/infra/http/response.go
@@ -28,6 +28,8 @@ func WriteHTTPErrorResponse(rw http.ResponseWriter, err error) {
 		writeErrorResponse(rw, http.StatusForbidden, err)
 	case errors.IsInvalidFormatError(err):
 		writeErrorResponse(rw, http.StatusBadRequest, err)
+	case errors.IsTooManyRequestError(err):
+		writeErrorResponse(rw, http.StatusTooManyRequests, err)
 	case errors.IsInvalidParameterError(err), errors.IsEncodingError(err):
 		writeErrorResponse(rw, http.StatusUnprocessableEntity, err)
 	case errors.IsHashicorpVaultError(err), errors.IsAKVError(err), errors.IsDependencyFailureError(err), errors.IsAWSError(err), errors.IsPostgresError(err):