+++ CHANGELOG.md
@@ -2,12 +2,13 @@
 
 ## v21.12.2 (Unreleased)
 ### 🆕 Features
-* Support for OIDC token custom claims `AUTH_OIDC_CUSTOM_CLAIMS` 
-* Support for OIDC token custom permission path `AUTH_OIDC_PERMISSIONS_CLAIMS`.
+* Support for OIDC token custom claims `AUTH_OIDC_CUSTOM_CLAIMS` for tenant_id and permissions.
   
 ### 🛠 Bug fixes
 * Invalid authentication forwarded to downstream proxy nodes if QKM authentication is enabled.
 * Fixed AWS client issue preventing racing condition on key state transition changes. 
+* Fixed invalid request error on create Registry with empty payload
+* Fixed forwarding of vault service 429 errors.
 
 ## v21.12.1 (2021-12-20)
 ### 🛠 Bug fixes